# server/src/docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Scan APIs
  description: Scan APIs
  license:
    name: Github
    url: 'https://github.com/rupayaproject/scan'
...
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/accounts:
    get:
      tags:
        - Accounts
      summary: Get list accounts
      parameters:
        - name: page
          in: query
          required: false
          description: default = 1, maximum = 500
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/accounts/{hash}:
    get:
      tags:
        - Accounts
      summary: Get account detail
      parameters:
        - name: hash
          in: path
          required: true
          description: account address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/accounts/{hash}/mined:
    get:
      tags:
        - Accounts
      summary: Get list block create
      parameters:
        - name: hash
          in: path
          required: true
          description: account address
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
...